src/05_column_locator.py
"""
V_OmegaFold: Structural Column Finder
Purpose: Analyzes PDB coordinates to find "Architectural Anchors" (Residue 121 style).
Identifies high-stability nodes that dictate the final 3D shape.
"""

import Bio.PDB # Requires 'pip install biopython'
import numpy as np

def locate_structural_columns(pdb_file):
    print(f"[V_OmegaFold] Analyzing structural anchors in: {pdb_file}")
    
    parser = Bio.PDB.PDBParser(QUIET=True)
    structure = parser.get_structure('protein', pdb_file)
    
    columns = []
    
    # Algorithm: Identifying residues with the lowest B-factor (highest stability)
    # and strategic spatial positioning (The 121 Residue Anomaly Logic)
    for model in structure:
        for chain in model:
            for residue in chain:
                # We extract the 'B-factor' which in simulations indicates 
                # where the protein is physically 'stiff' (a Column)
                avg_b_factor = np.mean([atom.get_bfactor() for atom in residue])
                
                if avg_b_factor < 20.0: # Threshold for a structural anchor
                    res_id = residue.get_id()[1]
                    res_name = residue.get_resname()
                    columns.append(f"{res_name}-{res_id}")

    print(f"\n[!] ANALYSIS COMPLETE: {len(columns)} Structural Columns detected.")
    print(f"[+] Primary Anchors identified: {columns[:5]}...")
    print("[+] These residues act as the 'physical skeleton' of the fold.")

if __name__ == "__main__":
    # Example: Analyzing our synthesized warhead
    try:
        locate_structural_columns("outputs/V_OMEGA_Warhead_PTP1B.pdb")
    except FileNotFoundError:
        print("[!] Error: Run 01_esmfold_local.py first to generate the PDB file.")